requirements.txt
py-cord==2.4.0
python-dotenv==0.21.1
requests==2.28.2